[refactor] Separate mem plugin
